test: move SiloRuntime cap tests to test_silo_runtime.cc
The two tests added in 3420eedc (TryRegisterCurrentThreadFailsGracefullyAtCap
and TryRegisterCurrentThreadIsIdempotent) exercise the Mako-side
SiloRuntime::try_register_current_thread API. They were in
test_masstree_multi_instance.cc only because that's where I was
working at the time — but conceptually they belong with the
SiloRuntime tests, not with the Masstree-context tests.

Why it matters: Masstree itself has no thread-count cap. The 512
limit is a SiloRuntime/Mako concept that surfaces *through*
Masstree only because this project's Mako-side simple_threadinfo
shim routes Masstree's allocations through Mako's RCU/coreid layer.
Keeping these tests under the "masstree multi-instance" header
created the misleading impression that the cap is a Masstree
limitation.

Moved verbatim into test_silo_runtime.cc under a new
SiloRuntimeThreadCap test suite. test_masstree_multi_instance.cc
keeps a short comment pointing at where they live now and drops
the SiloRuntime / core / macros includes it pulled in only for them.

Verified: both binaries build, test_silo_runtime is 10 cases
(8 existing + 2 moved), test_masstree_multi_instance is back to 9.
