+++ src/masstree/tests/test_masstree_multi_instance.cc
@@ -15,10 +15,7 @@
 
 #include "masstree/masstree_context.h"
 #include "masstree/kvthread.hh"
-#include "mako/core.h"
-#include "mako/macros.h"
 #include "mako/masstree_btree.h"
-#include "mako/silo_runtime.h"
 #include "mako/varkey.h"
 
 import std;
@@ -640,72 +637,10 @@ TEST_F(MasstreeMultiInstanceTest, RebindAcrossContextsKeepsBothConsistent) {
     }
 }
 
-// Test (Finding 6): explicit thread registration fails gracefully
-// when the per-runtime core-id pool is exhausted, instead of
-// aborting the process.
-//
-// Creates a FRESH SiloRuntime so the test doesn't poison the
-// process-wide default runtime's pool. Spawns N > NMAXCORES
-// worker threads, each of which calls try_register_current_thread
-// once and records success/failure. After join, expect exactly
-// NMAXCORES successes and the remainder failures.
-//
-// This is the safe alternative to the abort path documented in
-// docs/masstree-sanitizer-findings.md Finding 6.
-TEST_F(MasstreeMultiInstanceTest, TryRegisterCurrentThreadFailsGracefullyAtCap) {
-    rusty::Arc<SiloRuntime> rt_arc = SiloRuntime::Create();
-    SiloRuntime* rt = rt_arc.as_ptr();
-    ASSERT_NE(rt, nullptr);
-    EXPECT_EQ(rt->core_count(), 0u);
-
-    constexpr unsigned kOverhead = 8;  // a few extra threads past the cap
-    // Force NMaxCores into a real local — taking its address through
-    // gtest's CmpHelperEQ would otherwise need an out-of-class
-    // definition we don't have.
-    const unsigned cap = SiloRuntime::NMaxCores;
-    const unsigned kThreads = cap + kOverhead;
-
-    std::atomic<unsigned> registered{0};
-    std::atomic<unsigned> rejected{0};
-
-    std::vector<std::thread> workers;
-    workers.reserve(kThreads);
-    for (unsigned i = 0; i < kThreads; ++i) {
-        workers.emplace_back([rt, &registered, &rejected]() {
-            if (rt->try_register_current_thread()) {
-                ++registered;
-            } else {
-                ++rejected;
-            }
-        });
-    }
-    for (auto& t : workers) t.join();
-
-    EXPECT_EQ(registered.load(), cap);
-    EXPECT_EQ(rejected.load(),   kOverhead);
-}
-
-// Test (Finding 6, idempotency): try_register_current_thread is
-// a no-op (returns true without consuming a slot) for a thread
-// that already registered to the same runtime.
-TEST_F(MasstreeMultiInstanceTest, TryRegisterCurrentThreadIsIdempotent) {
-    rusty::Arc<SiloRuntime> rt_arc = SiloRuntime::Create();
-    SiloRuntime* rt = rt_arc.as_ptr();
-    ASSERT_NE(rt, nullptr);
-
-    std::thread t([rt]() {
-        EXPECT_TRUE(rt->try_register_current_thread());
-        const unsigned core_count_after_first = rt->core_count();
-        EXPECT_EQ(core_count_after_first, 1u);
-
-        // Subsequent calls must not consume another slot.
-        for (int i = 0; i < 10; ++i) {
-            EXPECT_TRUE(rt->try_register_current_thread());
-        }
-        EXPECT_EQ(rt->core_count(), core_count_after_first);
-    });
-    t.join();
-}
+// (Tests for SiloRuntime::try_register_current_thread — the Mako-side
+// fix for sanitizer-findings.md Finding 6 — live in test_silo_runtime.cc.
+// They belong with the SiloRuntime tests, not here, because the cap
+// is a SiloRuntime concept, not a Masstree concept.)
 
 // Test 9: Context teardown / leak check.
 //
